Each content-type format should define it's own encoding
There is a potential conflict between the existing, no-MUST/SHOULD/MAY
text in "The HTTP Request" section that says all GET requests should
be base64 and a later section that says it's true (MUST) for the
wireformat (which makes sense).  But I don't think we need to
require (or suggest) that all content-types have to use this encoding
format, as it may be unnecessary for some simple future version of
encoding.  Instead, let each content-type definition specify how
encoding must happen.