Security: update text about DNS server choice
The current text is difficult to understand, so replace it with
something clearer.

This is in connection to the following mail:
        https://www.ietf.org/mail-archive/web/doh/current/msg00543.html

I suppose that most computer users won't know what a DNS is and won't configure
a DNS API server themselves. Therefore "explicit configuration" in practice
means a "preconfigured DNS API server(s)".