Selection of DNS API Server: make it reference to "Security Considerations"
Both section "Security Considerations" and section "Selection of DNS API
Server" specify which server the client may use but in a different way:

	[...] the client MUST
	establish that the HTTP request URI is a trusted service for the DOH
	query, in other words, a DNS API client MUST only use a DNS API server
	that is configured as trustworthy.

vs

	A client MUST NOT use arbitrary DNS API servers. Instead, a client MUST
	only use DNS API servers specified using mechanisms such as explicit
	configuration.

Such discrepancy was confusing.

If there should be a section "Selection of DNS API Server", the text
there should be just a repetition of what is said in "Security
considerations".